third_party/WebKit/LayoutTests/inspector/console/exception-objects-expected.txt

CONSOLE ERROR: line 7: Uncaught Error: error_text
CONSOLE ERROR: line 11: Uncaught [object Object]
CONSOLE ERROR: line 15: Uncaught 42
CONSOLE ERROR: line 19: Uncaught (in promise) Error: promise_error
CONSOLE ERROR: line 19: Uncaught (in promise) Error: promise_error
CONSOLE ERROR: line 23: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 23: Uncaught (in promise) #<Object>
Tests that expressions have thrown objects.

setTimeout(throwError, 0); undefined
undefined
foo.js:3 Uncaught Error: error_text
    at throwError (foo.js:3)
throwError @ foo.js:3
setTimeout (async)
(anonymous) @ VM:1
throwError();
foo.js:3 Uncaught Error: error_text
    at throwError (foo.js:3)
    at <anonymous>:1:1
throwError @ foo.js:3
(anonymous) @ VM:1
setTimeout(throwObject, 0); undefined
undefined
foo.js:7 Uncaught {a: 42}
throwObject @ foo.js:7
setTimeout (async)
(anonymous) @ VM:1
throwObject();
foo.js:7 Uncaught {a: 42}
throwObject @ foo.js:7
(anonymous) @ VM:1
setTimeout(throwNumber, 0); undefined
undefined
foo.js:11 Uncaught 42
throwNumber @ foo.js:11
setTimeout (async)
(anonymous) @ VM:1
throwNumber();
foo.js:11 Uncaught 42
throwNumber @ foo.js:11
(anonymous) @ VM:1
setTimeout(rejectWithError, 0); undefined
undefined
foo.js:15 Uncaught (in promise) Error: promise_error
    at rejectWithError (foo.js:15)
rejectWithError @ foo.js:15
setTimeout (async)
(anonymous) @ VM:1
rejectWithError();
undefined
foo.js:15 Uncaught (in promise) Error: promise_error
    at rejectWithError (foo.js:15)
    at <anonymous>:1:1
rejectWithError @ foo.js:15
(anonymous) @ VM:1
setTimeout(rejectWithObject, 0); undefined
undefined
foo.js:19 Uncaught (in promise) {b: 42}
rejectWithObject @ foo.js:19
setTimeout (async)
(anonymous) @ VM:1
rejectWithObject();
undefined
foo.js:19 Uncaught (in promise) {b: 42}
rejectWithObject @ foo.js:19
(anonymous) @ VM:1
